+++ mapboard/topology_manager/commands/__init__.py
@@ -1,3 +1,4 @@
+from .check_setup import assert_topology_setup, check_topology_setup
 from .clean_topology import clean_topology
 from .create_tables import create_tables
 from .edge_relations import rebuild_edge_relations, validate_edge_relations